decouple decorators from func
